requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.104.0
fastapi-cli==0.0.7
flatbuffers==25.2.10
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.8
ml_dtypes==0.5.1
numpy==1.26.0
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==24.2
pillow==9.4.0
protobuf==4.25.6
pycparser==2.22
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.27.0
threadpoolctl==3.6.0
typer==0.15.2
typing_extensions==4.12.2
uvicorn==0.23.2
watchfiles==1.0.4
websockets==15.0.1
aiofiles==23.1.0
